+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
-* allow share links in ´!$more´ view
+* Add follow section
 * Add more networks
+
+## 1.2.1 - 2016-03-12
+**Fixed**
+
+* Share links in post's ´!$more´ view
+* Minor bugs
 
 ## 1.2.0 - 2016-03-12
 **Added**